requirements.txt
rivescript==1.15.0
discord.py==2.1.0
